Shifted decimal serialization functions from utils to serialization
