+++ dataclasses_avroschema/serialization.py
@@ -2,7 +2,7 @@ import datetime
 import io
 import typing
 import uuid
-
+import decimal
 import fastavro
 
 DATETIME_STR_FORMAT = "%Y-%m-%dT%H:%M:%S%z"
@@ -61,6 +61,45 @@ def time_to_str(value: datetime.time) -> str:
     return value.strftime(TIME_STR_FORMAT)
 
 
+def decimal_to_str(value: decimal.Decimal, precision: int, scale: int = 0) -> str:
+    value_bytes = prepare_bytes_decimal(value, precision, scale)
+    return r'\u' + value_bytes.hex()
+
+# This is an almost complete copy of fastavro's _logical_writers_py.prepare_bytes_decimal
+# the only tweak is to pass in scale/precision directly instead of a schema
+# This is needed to properly serialize a default decimal.Decimal into the avro schema
+def prepare_bytes_decimal(data, precision, scale=0):
+    """Convert decimal.Decimal to bytes"""
+
+    if not isinstance(data, decimal.Decimal):
+        return data
+
+    sign, digits, exp = data.as_tuple()
+
+    if len(digits) > precision:
+        raise ValueError(
+            'The decimal precision is bigger than allowed by schema')
+
+    delta = exp + scale
+
+    if delta < 0:
+        raise ValueError(
+            'Scale provided in schema does not match the decimal')
+
+    unscaled_datum = 0
+    for digit in digits:
+        unscaled_datum = (unscaled_datum * 10) + digit
+
+    unscaled_datum = 10 ** delta * unscaled_datum
+
+    bytes_req = (unscaled_datum.bit_length() + 8) // 8
+
+    if sign:
+        unscaled_datum = -unscaled_datum
+
+    return unscaled_datum.to_bytes(bytes_req, byteorder='big', signed=True)
+
+
 def to_json(data: typing.Dict[str, typing.Any]) -> typing.Dict:
     json_data = {}
 
@@ -73,7 +112,7 @@ def to_json(data: typing.Dict[str, typing.Any]) -> typing.Dict:
             value = date_to_str(value)
         elif isinstance(value, datetime.time):
             value = time_to_str(value)
-        elif isinstance(value, uuid.UUID):
+        elif isinstance(value, uuid.UUID) or isinstance(value, decimal.Decimal):
             value = str(value)
 
         json_data[field] = value
